REFACTORING COMPLETE
So basically:

Each character we want to add to the game will be created in the form of a prefab that contains a CharacterInfo class and the scripts for each of their moves. These move scripts should probably be unified in some sort of array but for now this is easier to modify as we go.

The move scripts also must be named correctly to become available to the player. The format is simply charactername + movename. This also should change eventually.

The player objects read the CharacterInfo class off the prefab and toss it all into a CharacterInfoCollector, which then supplies things like the name of the character, its starting health movement speed, etc to the scripts that need them at runtime.

MovementController and AbilityController have been modularized so we can controll each character with one script. This means all the important stuff is being done in the character prefab and its associated AnimationContoller (which is pulled by CharacterInfoCollector) so that the player objects ony have to worry about sending messages when buttons are pressed.

Im sure i changed more shit but I cant remember at the moment.

Ready now for more work on the actual mechanics of this thing. Woop!